Create Python program to print table of a given number
In this program, the user is asked to enter the number, and the program prints the multiplication table of the input number using for loop. The loops run from 1 to 10 and the input number is multiplied by the loop counter in each step to display the steps of the multiplication table.