Create Python program to find armstrong number
First Step: Take input from the user
Second Step: Find the cube of each digit of entered number
Third Step: Add the cube of all the digits
Fourth Step: If the output of step 3 is equal to the entered number i.e. Step 1. Then the print entered number is Armstrong number.
Fifth Step: If the output of step 3 is equal to the entered number i.e. Step 1. Then print entered number is not an Armstrong number.